Feature/all (#48)
* Feat: Network recv, send 관련 Core 로직 추가

- Core::setupNewClient() 함수에 User 등록 로직 추가.
- mNetwork.RecvFromClient() 실패시 User 삭제 로직 추가.
- mNetwork.SendToClient() 실패시 User 삭제 로직 추가.
- IRC 메시지 확인 및 로직 수행을 위한 Core::processIRCMessage() 추가.

* Feat: PASS, NICK, USER 메시지 로직 추가

- Core::handleIRCMessage() 메시지 파싱 후, 알맞는 메시지 처리 함수 호출.
- PASS 처리 함수 추가.
- NICK 처리 함수 추가.
- USER 처리 함수 추가.
- USER 등록 완료 후, Welcome 메시지(RPL_WELCOME) 추가 필요.

* Feat: PING, PONG 메시지 및 Channel 클래스 추가

- PING, PONG 메시지 로직 추가.
- Write 이벤트가 발생할 때마다 Core::checkUserConnection() 에서 client 연결 상태 확인.
- JOIN 메시지 로직 구현 중.
- PING, PONG 메시지까지 테스트 완료.

* Feat: JOIN 메시지 처리 로직 추가

- 채널이 없는 경우, 새로운 채널 추가.
- 채널이 있는 경우의 로직 추가 필요.

* Feat: JOIN 메시지 로직 구현 완료

- 단일 사용자 테스트 완료.
- 다수 사용자 테스트 필요.

* Style: 로그 메시지 변경

* Feat: Part 메시지 처리 로직 추가

- irssi에서 PART 응답 받으면 assertion 발생하는 문제 있음.
- 추가적인 테스트 및 수정 필요.

* Modify: PING, PONG 메시지 형식 수정

* Feat: 기존 채널에 유저 추가

* Modify: Log monitor 색상 회색으로 변경

* Modify: 채널의 operator 표시되도록 수정

* Modify: PING 메시지 형식 수정

* Refactor: Core->IRCMessage 리팩토링 진행 중

* Refactor: IRCMessage 리팩토링

- parseMessage 구현 필요.

* Refactor: IRC 리팩토링

* Refactor: UserManager, ChannelManager 추가

- User 관리를 위한 정적 클래스 UserManager 추가.
- Channel 관리를 위한 정적 클래스 ChannelManager 추가.

* Style: 불필요한 공백 제거

* Delete: IRC::CheckClientConnection() 제거

* Modify: IRC 로직 수정

- PONG 메시지 받으면 유저 등록 및 Welcome 메시지 전달 되도록 수정.
- static 멤버 변수명 변경 -> sStatic.
- Welcome 메시지 추가(001, 002, 003, 004, 005).
- MOTD 응답 추가.

* Feat: PRIVMSG 메시지 로직 추가

- PART 메시지 버그 수정 필요.

* Fix: PART 메시지 버그 해결

* Refactor: JOIN 메시지 로직 추가

* Refactor: IRC function map

- 함수 호출을 맵으로 변경, N -> logN
- C style void casting -> C++ style void casting (static_cast)
- makefile -> exception, unwind table 제거 플래그 추가

* Fix: IRC::HandleMessage, Core::Run

- IRC::HandleMessage
- 문제: 함수포인터 맵에 키가 존재하지 않을 경우 쓰레기값에 접근함.
- 해결: 키가 존재하는지 확인 후 접근

- Core::Run
- 문제: 연결이 끊길 때, 채널 삭제 전 유저가 삭제되기 때문에 문제가 생김
- 해결: 채널내 유저 삭제 후, 유저 완전 삭제로 변경

* Fix: IRC::PRIVMSG, IRC::JOIN

- IRC::PRIVMSG
- 문제: 유저간 PRIVMSG 전송시, 무한 루프 발생.
- 해결: UserManager::GetUser(const std::String&)에서 it++ 추가.

- IRC::JOIN
- 문제: 채널에 입장시, 유저 목록에 한 명만 나오는 현상.
- 해결: Channel::GetAllUsersNickname()의 else 구문 수정
        Before: it->first
        After:  userNickname = it->first;

* Feat: KICK 메시지 로직 추가

- 채널 오퍼레이터만 특정 유저 추방 가능.
- 예외 처리 테스트 완료.
- 여러 유저 추방 테스트 완료.

* Feat: TOPIC 메시지 로직 추가

- 채널의 유저가 topic 변경 및 확인 가능.
- 현재는 채널의 오퍼레이터만 변경 가능하나,
  MODE 로직 추가 후, 해당 예외처리 로직 수정 필요.
- 테스트 완료.

* Feat: MODE 메시지 로직 추가

- MODE 메시지 형식.
  MODE | modestring | modeargument
- modestring은 +또는 -로 시작하며, '+'는 mode 추가, '-'는 mode 삭제.
- modestring의 종류
  1. i = 초대 전용 채널로 설정.
  2. t = 채널 관리자(오퍼레이터)만 채널의 토픽 설정 가능.
  3. k = 채널 입장을 위한 key(비밀번호) 설정.
  4. o = 특정 유저를 채널 관리자로 설정 및 해제.
  5. l = 채널에 최대 입장 가능한 유저 수 설정.
- 지원하지 않는 mode를 받은 경우 무시하도록 구현.
- 테스트 완료.
- MODE 구현 완료에 따른 JOIN, TOPIC 예외 처리 테스트 완료.

* Modify: 채널 대소문자 구별 제거

* Modify: QUIT 메시지 수신시, 해당 유저가 속한 채널에 알림

* Style: LOG 메시지 위치 변경

* Feat: NICK 메시지 로직 추가

- 기존에 존재하던 유저인 경우, NICK 메시지로 닉네임 변경 로직 가능하도록 구현.
- 새로운 유저인 경우 새로 등록됨.
- 닉네임 변경시, 유저가 속한 Channel.mUsers에 새로운 닉네임으로 추가되며,
  기존 닉네임은 삭제됨.
- 테스트 완료.

* Fix: NICK 메시지 버그 및 유령 채널 버그

- 처음 등록된 유저인 경우 NICK 메시지 응답하는 버그 해결.
- 유저 등록 및 삭제 로그 추가.
- 채널 오퍼레이터 등록 및 삭제 로그 추가.
- 유저가 QUIT 메시지로 종료시, 현재 채널들에서 나가지고 빈 채널인 경우 삭제.
- 채널 삭제 로그 추가.
- 비슷한 로직들 일관성 있도록 코드 수정.

* Refactor: 코드 리팩토링

* Feat: INVITE 메시지 로직 추가

- 초대 전용 채널인 경우 INVITE 된 유저만 입장 가능하도록 구현.
- 채널 나가면 다시 초대 받아야 들어올 수 있도록 구현.
- 예외 처리 테스트 완료.
- 종합 테스트 완료.

* Feat: status 명령 user channel 수 표시

* Modify: Mode <channel> b 무시하도록 수정

* Feat: Network::ReservedDisconnectClient(socket) 구현

- Session 구조체 bool 멤버 변수 isReservedDisconnect 추가.
- ReservedDisconnectClient(socket) 함수는 해당 socket 세션의
  isReservedDisconnect = true로 설정.
- RecvFromClient() 에서 isReservedDisconnect가 true인 경우, 생략됨.
- SendToClient() 에서 Send Buffer에 남아있는 데이터가 없고
  isReservedDisconnect가 true인 경우, 해당 소켓 연결 해제.

* Fix: KICK 당해도 invite 채널에 입장되는 버그 수정

* Feat: Earth animation

- IAnimation.hpp, Earth.hpp, Earth.cpp 추가
- Core Earth 멤버 추가
- Core Run() 메소드에 반영

* Modify: better Earth animation logic

- 애니메이션이 60ms 마다 업데이트 되도록 변경(Server Monitor 전체)
- IAnimation 인터페이스 함수 변경
- Earth.hpp, Earth.cpp 클래스 로직 단순화
- Core.hpp, Core.hpp 변경된 로직 적용
- Core::isTimePassed(ms, lastTime) 함수 추가

* Refactor: include header

* Fix: BSD-GDF color error

* Feat: monitor footer 추가

* Feat: invite의 target user가 없는 경우 생략 추가

* Feat: 다수의 MODE 명령어 적용 추가

* Feat: 유저가 속한 채널에서 삭제될 때, 채널 비면 해당 채널 삭제

* Style: WhiteCharGrayBG 색상 변경 \033[100;37m -> \033[47;37m

---------

+++ src/grc/irc/IRC.cpp
@@ -118,13 +118,25 @@ void IRC::PASS(const int32 IN socket,
         messageToReply.append(":Not enough parameters");
         messageToReply.append(CRLF);
         network.PushToSendBuffer(socket, messageToReply);
-        // 즉시 메시지 전송 후, client 연결 해제
-        network.SendToClient(socket);
-        network.DisconnectClient(socket);
+        network.ReserveDisconnectClient(socket);
+        network.ClearRecvBuffer(socket);
         UserManager::DeleteUser(socket);
         return;
     }
     const std::string& passwordInMessage = parameters[0];
+    // ERR_PASSWDMISMATCH
+    if (passwordInMessage != password)
+    {
+        messageToReply.append(ERR_PASSWDMISMATCH);
+        messageToReply.append(" ");
+        messageToReply.append(":Password incorrect");
+        messageToReply.append(CRLF);
+        network.PushToSendBuffer(socket, messageToReply);
+        network.ReserveDisconnectClient(socket);
+        network.ClearRecvBuffer(socket);
+        UserManager::DeleteUser(socket);
+        return;
+    }
     User& user = UserManager::GetUser(socket);
     // ERR_ALREADYREGISTERED
     if (user.IsAuthenticated())
@@ -134,20 +146,6 @@ void IRC::PASS(const int32 IN socket,
         messageToReply.append(":You may not reregister");
         messageToReply.append(CRLF);
         network.PushToSendBuffer(socket, messageToReply);
-        return;
-    }
-    // ERR_PASSWDMISMATCH
-    if (passwordInMessage != password)
-    {
-        messageToReply.append(ERR_PASSWDMISMATCH);
-        messageToReply.append(" ");
-        messageToReply.append(":Password incorrect");
-        messageToReply.append(CRLF);
-        network.PushToSendBuffer(socket, messageToReply);
-        // 즉시 메시지 전송 후, client 연결 해제
-        network.SendToClient(socket);
-        network.DisconnectClient(socket);
-        UserManager::DeleteUser(socket);
         return;
     }
     user.SetAuthenticated();
@@ -733,6 +731,12 @@ void IRC::MODE(const int32 IN socket,
         network.PushToSendBuffer(socket, messageToReply);
         return;
     }
+    const std::string& modeString = parameters[1];
+    // 지원하지 않는 옵션(밴 리스트 출력)인 경우 생략
+    if (modeString[0] == 'b')
+    {
+        return;
+    }
     // modestring이 있는데 채널의 오퍼레이터가 아닌 경우, ERR_CHANOPRIVSNEEDED 응답
     // ERR_CHANOPRIVSNEEDED
     if (ChannelManager::GetChannel(channelName).IsOperator(user.GetNickname()) == false)
@@ -749,13 +753,16 @@ void IRC::MODE(const int32 IN socket,
         return;
     }
     // MODE 설정 또는 해제
-    const std::string& modeString = parameters[1];
     std::vector<std::string> modeArgument;
     if (parameters.size() >= 3)
     {
-        modeArgument = split(parameters[2], " ");
+        for (std::size_t i = 2; i < parameters.size(); i++)
+        {
+            modeArgument.push_back(parameters[i]);
+        }
     }
     std::string resultModeString("");
+    std::string resultArgumentString("");
     std::size_t modeStringIndex = 0;
     std::size_t modeArgumentIndex = 0;
     bool isAddType = false;
@@ -779,6 +786,11 @@ void IRC::MODE(const int32 IN socket,
                 {
                     break;
                 }
+                if (channel.IsUserExist(modeArgument[modeArgumentIndex]) == false)
+                {
+                    modeArgumentIndex++;
+                    break;
+                }
                 channel.AddOperator(modeArgument[modeArgumentIndex],
                                UserManager::GetUser(modeArgument[modeArgumentIndex]));
             }
@@ -788,10 +800,17 @@ void IRC::MODE(const int32 IN socket,
                 {
                     break;
                 }
+                if (channel.IsUserExist(modeArgument[modeArgumentIndex]) == false)
+                {
+                    modeArgumentIndex++;
+                    break;
+                }
                 channel.DeleteOperator(modeArgument[modeArgumentIndex]);
             }
-            modeArgumentIndex++;
             resultModeString.append("o");
+            resultArgumentString.append(modeArgument[modeArgumentIndex]);
+            resultArgumentString.append(" ");
+            modeArgumentIndex++;
             break;
         // 토픽을 오퍼레이터만 변경할 수 있도록 함.
         case 't':
@@ -827,6 +846,8 @@ void IRC::MODE(const int32 IN socket,
                 }
                 channel.SetLimitedMaxUserCount();
                 channel.SetMaxUserCount(std::atoi(modeArgument[modeArgumentIndex].c_str()));
+                resultArgumentString.append(modeArgument[modeArgumentIndex]);
+                resultArgumentString.append(" ");
                 modeArgumentIndex++;
             }
             else
@@ -846,6 +867,8 @@ void IRC::MODE(const int32 IN socket,
                 }
                 channel.SetKeyRequired();
                 channel.SetKey(modeArgument[modeArgumentIndex]);
+                resultArgumentString.append(modeArgument[modeArgumentIndex]);
+                resultArgumentString.append(" ");
                 modeArgumentIndex++;
             }
             else
@@ -874,11 +897,7 @@ void IRC::MODE(const int32 IN socket,
     messageToReply.append(" ");
     messageToReply.append(resultModeString);
     messageToReply.append(" ");
-    for (std::size_t i = 0; i < modeArgument.size(); i++)
-    {
-        messageToReply.append(modeArgument[i]);
-        messageToReply.append(" ");
-    }
+    messageToReply.append(resultArgumentString);
     messageToReply.pop_back(); // 맨 끝의 공백 문자 제거
     messageToReply.append(CRLF);
     // 보낼 메시지가 없는 경우 생략
@@ -1060,6 +1079,10 @@ void IRC::INVITE(const int32 IN socket,
         return;
     }
     const std::string& targetUser = parameters[0];
+    if (UserManager::IsUserExist(targetUser) == false)
+    {
+        return;
+    }
     const std::string& channelName = parameters[1];
     // ERR_NOSUCHCHANNEL
     if ((channelName[0] != '#' && channelName[0] != '&') || channelName.find(' ') != std::string::npos
@@ -1264,7 +1287,7 @@ void IRC::KICK(const int32 IN socket,
     }
     channel.DeleteUser(targetUser);
     channel.DeleteOperator(targetUser);
-    channel.DeleteInvitedUser(user.GetNickname());
+    channel.DeleteInvitedUser(targetUser);
     ChannelManager::CheckIsEmptyChannelAndDelete(channel);
 }
